requirements.txt
pypdf
pandas
python-dotenv
matplotlib
anthropic
google-genai
jsonschema
